requirements.txt
requests==2.32.3
urllib3<2
openai==1.57.0
PyYAML==6.0.2

